requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/ # https://github.com/microsoft/onnxruntime/issues/21684
anyio==4.9.0
atomics==1.0.3
black==25.1.0
boltons==24.0.0
boto3==1.38.30
botocore==1.38.30
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
coloredlogs==15.0.1
conformer==0.3.2
cryptography==45.0.3
diffusers==0.29.0
easydict==1.13
einops==0.8.1
environs==14.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpeg==1.4
filelock==3.17.0
frozendict==2.4.6
fsspec==2024.12.0
gdown==5.1.0
gmpy2==2.2.1
gunicorn==23.0.0
h11==0.16.0
huggingface-hub==0.32.4
humanfriendly==10.0
humanize==4.12.3
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
inflect==7.3.1
Jinja2==3.1.5
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
librosa==0.10.2
lightning==2.2.4
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.7.5
mpmath==1.3.0
multiprocessing-logging==0.3.4
networkx==3.1
ninja==1.11.1.3
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.18.0
onnxruntime-gpu==1.16.3
openai-whisper
packaging==24.2
Pillow==11.0.0
pip==24.3.1
pluggy==1.5.0
plumbum==1.9.0
prometheus_client==0.22.1
protobuf==4.25
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.7.0
pyOpenSSL==25.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyworld==0.3.4
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rpyc==6.0.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
s3transfer==0.13.0
schema==0.7.7
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.46.2
sympy==1.13.1
tensorrt-cu12==10.7.0
tensorrt-cu12-bindings==10.7.0
tensorrt-cu12-libs==10.7.0
tiktoken==0.9.0
toolz==1.0.0
typing_extensions==4.14.0
urllib3==2.3.0
uvicorn==0.19.0
uvloop==0.21.0
websockets==11.0.3
WeTextProcessing==1.0.3
wget==3.2
wheel==0.45.1
zstandard==0.23.0
